EPMRPP-92607 || Code Review fix - 1
