+++ app/src/controllers/user/selectors.js
@@ -16,7 +16,6 @@
 
 import { ADMINISTRATOR } from 'common/constants/accountRoles';
 import { createSelector } from 'reselect';
-import { MANAGER } from 'common/constants/projectRoles';
 import { START_TIME_FORMAT_ABSOLUTE, START_TIME_FORMAT_RELATIVE } from './constants';
 
 const userSelector = (state) => state.user || {};
@@ -38,34 +37,6 @@ export const assignedProjectsSelector = (state) => userInfoSelector(state).assig
 export const assignedOrganizationsSelector = (state) =>
   userInfoSelector(state).assignedOrganizations || {};
 export const userAccountRoleSelector = (state) => userInfoSelector(state).userRole || '';
-export const activeProjectRoleSelector = createSelector(
-  activeProjectSelector,
-  assignedProjectsSelector,
-  (activeProject, assignedProjects) => {
-    const { projectSlug } = activeProject;
-    const assignedProject = assignedProjects[projectSlug];
-    return assignedProject?.projectRole;
-  },
-);
-const activeOrganizationRoleSelector = createSelector(
-  activeProjectSelector,
-  assignedOrganizationsSelector,
-  (activeProject, assignedOrganizations) => {
-    const { organizationSlug } = activeProject;
-    const assignedOrganization = assignedOrganizations[organizationSlug];
-    return assignedOrganization?.organizationRole;
-  },
-);
-export const userRolesSelector = createSelector(
-  userAccountRoleSelector,
-  activeOrganizationRoleSelector,
-  activeProjectRoleSelector,
-  (userRole, organizationRole, projectRole) => ({
-    userRole,
-    organizationRole,
-    projectRole,
-  }),
-);
 export const isAdminSelector = (state) => userInfoSelector(state).userRole === ADMINISTRATOR;
 
 export const availableProjectsSelector = createSelector(
@@ -106,43 +77,3 @@ export const availableProjectsSelector = createSelector(
 export const apiKeysSelector = (state) => userSelector(state).apiKeys || [];
 
 export const activeProjectKeySelector = (state) => userSelector(state).activeProjectKey;
-
-export const createUserAssignedSelector = (projectSlug, organizationSlug) =>
-  createSelector(
-    userRolesSelector,
-    assignedOrganizationsSelector,
-    assignedProjectsSelector,
-    (userRoles, assignedOrganizations, assignedProjects) => {
-      const { userRole, organizationRole } = userRoles;
-      const isAdmin = userRole === ADMINISTRATOR;
-      const isManager = organizationRole === MANAGER;
-      let isAssignedToTargetOrganization = false;
-
-      if (organizationSlug) {
-        isAssignedToTargetOrganization = organizationSlug in assignedOrganizations;
-      } else {
-        const organizationId = assignedProjects[projectSlug]?.organizationId || '';
-        isAssignedToTargetOrganization = Object.keys(assignedOrganizations).some(
-          (key) => assignedOrganizations[key]?.organizationId === organizationId,
-        );
-      }
-
-      const isAssignedToTargetProject =
-        projectSlug && projectSlug in assignedProjects && isAssignedToTargetOrganization;
-
-      const assignmentNotRequired = isAdmin || (isManager && isAssignedToTargetOrganization);
-
-      const hasPermission = isAssignedToTargetProject || assignmentNotRequired;
-
-      const assignedProjectKey = assignedProjects?.[projectSlug]?.projectKey;
-
-      return {
-        isAdmin,
-        hasPermission,
-        assignedProjectKey,
-        assignmentNotRequired,
-        isAssignedToTargetProject,
-        isAssignedToTargetOrganization,
-      };
-    },
-  );
