+++ app/src/pages/common/membersPage/membersGrid/projectRole/projectRole.jsx
@@ -25,8 +25,8 @@ import { fetch } from 'common/utils';
 import { URLS } from 'common/urls';
 import { InputDropdown } from 'components/inputs/inputDropdown';
 import { canChangeUserRole } from 'common/utils/permissions';
-import { urlProjectSlugSelector } from 'controllers/pages';
-import { userIdSelector, userRolesSelector } from 'controllers/user';
+import { urlProjectSlugSelector, userRolesSelector } from 'controllers/pages';
+import { userIdSelector } from 'controllers/user';
 import { MEMBERS_PAGE_EVENTS } from 'components/main/analytics/events';
 import { ROLES_MAP } from 'common/constants/projectRoles';
 import { ADMINISTRATOR } from 'common/constants/accountRoles';
