+++ app/src/pages/common/membersPage/membersGrid/unassignButton/unassignButton.jsx
@@ -26,9 +26,9 @@ import { fetch } from 'common/utils';
 import { URLS } from 'common/urls';
 import { COMMON_LOCALE_KEYS } from 'common/constants/localization';
 import { canAssignUnassignInternalUser } from 'common/utils/permissions';
-import { urlProjectSlugSelector } from 'controllers/pages';
+import { urlProjectSlugSelector, userRolesSelector } from 'controllers/pages';
 import { userRolesType } from 'common/constants/projectRoles';
-import { userIdSelector, assignedProjectsSelector, userRolesSelector } from 'controllers/user';
+import { userIdSelector, assignedProjectsSelector } from 'controllers/user';
 import { showNotification, NOTIFICATION_TYPES } from 'controllers/notification';
 import { MEMBERS_PAGE_EVENTS } from 'components/main/analytics/events';
 import UnassignIcon from 'common/img/unassign-inline.svg';
