+++ app/src/pages/inside/common/launchFiltersToolbar/launchFiltersToolbar.jsx
@@ -39,7 +39,7 @@ import AddFilterIcon from 'common/img/add-filter-inline.svg';
 import { LAUNCHES_PAGE_EVENTS } from 'components/main/analytics/events';
 import { canWorkWithFilters } from 'common/utils/permissions';
 import { userRolesType } from 'common/constants/projectRoles';
-import { userRolesSelector } from 'controllers/user';
+import { userRolesSelector } from 'controllers/pages';
 import { FilterList } from './filterList';
 import { FiltersActionBar } from './filtersActionBar';
 import { ExpandToggler } from './expandToggler';
