+++ app/src/pages/organization/organizationProjectsPage/projectsListTable/projectName/projectName.jsx
@@ -22,8 +22,8 @@ import { useDispatch, useSelector } from 'react-redux';
 import { ADMIN_PROJECTS_PAGE_EVENTS } from 'components/main/analytics/events';
 import { SCREEN_XS_MAX_MEDIA } from 'common/constants/screenSizeVariables';
 import { navigateToProjectAction } from 'controllers/organizations/projects';
-import { createUserAssignedSelector, setActiveProjectKeyAction } from 'controllers/user';
-import { PROJECT_PAGE } from 'controllers/pages';
+import { setActiveProjectKeyAction } from 'controllers/user';
+import { createUserAssignedSelector, PROJECT_PAGE } from 'controllers/pages';
 import styles from './projectName.scss';
 
 const cx = classNames.bind(styles);
