+++ app/src/pages/organization/projectTeamPage/projectTeamListTable/projectTeamListTable.jsx
@@ -19,11 +19,11 @@ import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
 import { useMemo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { activeProjectKeySelector, userRolesSelector } from 'controllers/user';
+import { activeProjectKeySelector } from 'controllers/user';
 import { AbsRelTime } from 'components/main/absRelTime';
 import { MeatballMenuIcon, Popover, Table } from '@reportportal/ui-kit';
 import { UserAvatar } from 'pages/inside/common/userAvatar';
-import { urlOrganizationAndProjectSelector } from 'controllers/pages';
+import { urlOrganizationAndProjectSelector, userRolesSelector } from 'controllers/pages';
 import { SORTING_ASC, withSortingURL } from 'controllers/sorting';
 import { DEFAULT_SORT_COLUMN } from 'controllers/members/constants';
 import { fetchMembersAction } from 'controllers/members';
